256qam: Fix incorrect LLR calculation
The optimized demapper was not used. Further, it lacked a correct
implementation. Both issues are fixed now.

Some old and primitive benchmark code was removed. `chrono` is not used
anymore.

The generic demapper methods are available directly via the public API
now. This should make it easier to benchmark them.